# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SU'UD Job Portal API",
        "description": "Complete API documentation for the SU'UD Job Portal platform. This API provides comprehensive job management, user authentication, application tracking, and administrative features for a modern job portal platform.",
        "contact": {
            "name": "SU'UD Development Team",
            "email": "admin@suud.sa"
        },
        "license": {
            "name": "Private License",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.suud.sa",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Admin Dashboard Overview",
                "description": "Get comprehensive dashboard statistics including users, jobs, applications, and companies data",
                "operationId": "049dace0f34f5f9cd2c0d7ee98daf583",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "stats": {
                                                    "properties": {
                                                        "users": {
                                                            "properties": {
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 1250
                                                                },
                                                                "employees": {
                                                                    "type": "integer",
                                                                    "example": 980
                                                                },
                                                                "employers": {
                                                                    "type": "integer",
                                                                    "example": 265
                                                                },
                                                                "admins": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "active": {
                                                                    "type": "integer",
                                                                    "example": 1198
                                                                },
                                                                "inactive": {
                                                                    "type": "integer",
                                                                    "example": 52
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "jobs": {
                                                            "properties": {
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 450
                                                                },
                                                                "active": {
                                                                    "type": "integer",
                                                                    "example": 280
                                                                },
                                                                "pending": {
                                                                    "type": "integer",
                                                                    "example": 45
                                                                },
                                                                "declined": {
                                                                    "type": "integer",
                                                                    "example": 80
                                                                },
                                                                "closed": {
                                                                    "type": "integer",
                                                                    "example": 45
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "applications": {
                                                            "properties": {
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 3240
                                                                },
                                                                "pending": {
                                                                    "type": "integer",
                                                                    "example": 456
                                                                },
                                                                "reviewed": {
                                                                    "type": "integer",
                                                                    "example": 1234
                                                                },
                                                                "accepted": {
                                                                    "type": "integer",
                                                                    "example": 890
                                                                },
                                                                "rejected": {
                                                                    "type": "integer",
                                                                    "example": 660
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - Admin role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}